[libromdata] CIAReader: `sz_dec` should be `size_t`, not `unsigned int`.
This fixes an MSVC 2017 15.7 warning:
src\libromdata\disc\ciareader.cpp(369): warning C4267: 'initializing':
conversion from 'size_t' to 'unsigned int', possible loss of data

+++ src/libromdata/disc/CIAReader.cpp
@@ -366,7 +366,7 @@ size_t CIAReader::read(void *ptr, size_t size)
 		m_lastError = EIO;
 		return 0;
 	}
-	unsigned int sz_dec = d->cipher->decrypt(static_cast<uint8_t*>(ptr), size);
+	size_t sz_dec = d->cipher->decrypt(static_cast<uint8_t*>(ptr), size);
 	if (sz_dec != size) {
 		// decrypt() failed.
 		m_lastError = EIO;
